+++ apps/www/src/app/(main)/(docs)/components/charts/area-chart/examples/area-chart-1.tsx
@@ -2,7 +2,7 @@
 
 "use client";
 
-import {  Card, CardContent, CardHeader } from "@jamsr-ui/react";
+import { Card, CardContent, CardHeader } from "@jamsr-ui/react";
 import {
   Area,
   AreaChart,