+++ apps/www/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web
 
+## 0.1.198
+
+### Patch Changes
+
+- fix react package
+- Updated dependencies
+  - @jamsr-ui/react@0.0.216
+
 ## 0.1.197
 
 ### Patch Changes